Move the entire crate to v1.
